skills/Youtube-clipper-skill/requirements.txt
yt-dlp
pysrt
python-dotenv
curl_cffi
